squashed commits due to initially commiting files too large
commit messages below

work in progress where circuit using input size 25 compiles proves verifies

- need to implement in browser end to end and repeat for input size 1400
- did ptau ceremony for pot19 -- very large

working end-to-end with input size 25

working end to end with input size 1400

- cannot commit pot19 ptau files to github due to size
- this interface is ugly but shows the proof of concept for integration into other projects

cannot commit large files generated by circom
