+++ package.json
@@ -11,13 +11,15 @@
     "@types/react": "^17.0.0",
     "@types/react-dom": "^17.0.0",
     "circomlib": "^2.0.5",
+    "circomlibjs": "^0.1.7",
     "cors": "^2.8.5",
     "express": "^4.17.1",
     "ffjavascript": "^0.2.35",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-scripts": "4.0.3",
-    "snarkjs": "^0.3.59",
+    "readline": "^1.3.0",
+    "snarkjs": "0.6.10",
     "styled-components": "^5.2.1",
     "typescript": "^4.1.2",
     "web-vitals": "^1.0.1"
