+++ src/App.tsx
@@ -1,14 +1,22 @@
 import React, { useState } from "react";
 import "./App.css";
 
-// I got rid of rimble-ui because of dependency issues
-//import { Field, Input, Text, Button, Link } from "rimble-ui";
-
 const snarkjs = require("snarkjs");
-const fs = require("fs");
 const { buildBabyjub, buildMimc7, buildEddsa } = require("circomlibjs");
 
-const MAX_JSON_SIZE = 1000; // TODO: increase the size after testing (slower)
+// Use circuit files with size matching MAX_JSON_SIZE
+const wasmFile = "http://localhost:8000/circuit_25_js/circuit_25.wasm";
+const zkeyFile = "http://localhost:8000/circuit_25.zkey";
+const verificationKey = "http://localhost:8000/verification_key_25.json";
+// const wasmFile = "http://localhost:8000/circuit_1400_js/circuit_1400.wasm";
+// const zkeyFile = "http://localhost:8000/circuit_1400.zkey";
+// const verificationKey = "http://localhost:8000/verification_key_1400.json";
+
+// Update this constant to match the circuit version
+// small is for testing (faster prove/verify)
+const MAX_JSON_SIZE = 25;
+// const MAX_JSON_SIZE = 1400;
+
 
 // utilities
 export type Ascii = number;
@@ -26,65 +34,37 @@ const isValidJson = (jsonInput) => {
 }
 
 
-const makeProof = async (_proofInput: any, _wasm: string, _zkey: string) => {
-	const { proof, publicSignals } = await snarkjs.groth16.fullProve(_proofInput, _wasm, _zkey);
-	return { proof, publicSignals };
-};
-
-const verifyProof = async (_verificationkey: string, signals: any, proof: any) => {
-	const vkey = await fetch(_verificationkey).then(function (res) {
-		return res.json();
-	});
-	const res = await snarkjs.groth16.verify(vkey, signals, proof);
-	return res;
-};
-
 function App() {
-	const [a, setA] = useState("3");
-	const [b, setB] = useState("11");
 	const [jsonInput, setJsonInput] = useState(`{"name": "aberke"}`);
+	const [paddedJsonAscii, setPaddedJsonAscii] = useState<any | null>(null);
 
 	const [proof, setProof] = useState("");
 	const [signals, setSignals] = useState("");
 	const [isValid, setIsValid] = useState(false);
 
-	let wasmFile = "http://localhost:8000/circuit.wasm";
-	let zkeyFile = "http://localhost:8000/circuit_final.zkey";
-	let verificationKey = "http://localhost:8000/verification_key.json";
+	const [proofInput, setProofInput] = useState<any | null>(null);
 
-	const runProofs = () => {
-		console.log(b.length);
-		if (a.length == 0 || b.length == 0) {
+	const generateProof = async () => {
+		if (!proofInput) {
+			console.error('no valid proof input');
 			return;
 		}
-		let proofInput = { a, b };
-		console.log(proofInput);
+		console.log('generating proof...')
+		const { proof, publicSignals } = await snarkjs.groth16.fullProve(proofInput, wasmFile, zkeyFile);
+		console.log('generated proof')
+		setProof(JSON.stringify(proof, null, 2));
+		setSignals(JSON.stringify(publicSignals, null, 2));
+	};
 
-		makeProof(proofInput, wasmFile, zkeyFile).then(({ proof: _proof, publicSignals: _signals }) => {
-			setProof(JSON.stringify(_proof, null, 2));
-			setSignals(JSON.stringify(_signals, null, 2));
-			verifyProof(verificationKey, _signals, _proof).then((_isValid) => {
-				setIsValid(_isValid);
-			});
+	const handleVerifyProof = async () => {
+		const vkey = await fetch(verificationKey).then(function (res) {
+			return res.json();
 		});
-	};
-	const changeA = (e) => {
-		setA(e.target.value);
-	};
-	const changeB = (e) => {
-		setB(e.target.value);
+		const isValid = await snarkjs.groth16.verify(vkey, JSON.parse(signals), JSON.parse(proof));
+		setIsValid(isValid);
 	};
 
-
-
-	const handleJsonInput = async () => {
-		console.log('handleJsonInput')
-		// Transform JSON input to a string of ascii bytes
-		// Check if this is valid JSON (though we do not need it as an object for now)
-		if (!isValidJson(jsonInput)) {
-			console.log('invalid json input:', jsonInput); // TODO: error handling
-			return false;
-		}
+	const jsonToPaddedAsciiArray = (jsonInput: string) => {
 		// Convert to object and back to string in order to have consistent representation
 		// e.g. consistent key sorting, removed whitespace
 		const parsedJson = JSON.stringify(JSON.parse(jsonInput));
@@ -92,29 +72,40 @@ function App() {
 		// Note: Doesn't matter if this is array of integers or strings of integers
 		// i.e. {"test":[1,0]} same as {"test":["1", "0"]}
 		const jsonAsciiArray = toAsciiArray(parsedJson);
-		console.log('jsonAsciiArray')
-		console.log(jsonAsciiArray);
 		// Pad the ascii array to the predetermined size.
 		// This is necessary because a predetermined size must be used in compiled circuit.
 		// First check it does not exceed this size.
 		if (jsonAsciiArray.length > MAX_JSON_SIZE) {
 			console.error('Cannot handle json: too large. Max size:', MAX_JSON_SIZE, 'Json array size:', jsonAsciiArray.length);
-			return false;
+			return;
 		}
 		const jsonSize = jsonAsciiArray.length; // Size before padding
 		const paddingLength = MAX_JSON_SIZE - jsonSize;
-		const paddedInput = jsonAsciiArray.concat(Array(paddingLength).fill(0));
-		console.log('paddedInput')
-		console.log(paddedInput)
+		return jsonAsciiArray.concat(Array(paddingLength).fill(0));
+	};
 
+	const handleInput = async () => {
+		// Transform JSON input to a string of ascii bytes
+		// Check if this is valid JSON (though we do not need it as an object for now)
+		if (!isValidJson(jsonInput)) {
+			console.log('invalid json input:', jsonInput); // TODO: error handling
+			return false;
+		}
+		const paddedJsonAsciiArray = jsonToPaddedAsciiArray(jsonInput);
+		if (!paddedJsonAsciiArray) {
+			return;
+		}
+		setPaddedJsonAscii(paddedJsonAsciiArray);
+
+		// Handle hashing and signature
 		const mimc7 = await buildMimc7();
 		const babyJub = await buildBabyjub();
 		const F = babyJub.F;
 
 		const k = 1;
-		const hash = mimc7.multiHash(paddedInput, k);
-
+		const hash = mimc7.multiHash(paddedJsonAsciiArray, k);
 		// handle signature
+		// TODO: Key management. This is just an example
 		const eddsa = await buildEddsa();
 		console.log('generating Eddsa (sk, pk) key pair')
 		const sk = Buffer.from("1".toString().padStart(64, "0"), "hex");
@@ -124,7 +115,7 @@ function App() {
 		const signature = eddsa.signMiMC(sk, hash);
 
 		const inputs = {
-			json: paddedInput,
+			json: paddedJsonAsciiArray,
 			expected_hash: BigInt(F.toObject(hash)).toString(),
 			pubkey: [
 				BigInt(F.toObject(pk[0])).toString(),
@@ -135,11 +126,8 @@ function App() {
 			signature_S: BigInt(signature.S).toString(),
 		};
 		console.log('inputs')
-		console.log(JSON.stringify(inputs))
-	}
-
-	const handleSignature = () => {
-		// TODO: move things here
+		console.log(JSON.stringify(inputs));
+		setProofInput(inputs);
 	}
 
 	const changeJsonInput = (e) => {
@@ -150,37 +138,34 @@ function App() {
 		<div>
 			<header className="App-header">
 				<p>
-					Hash and sign JSON with MIMC - testing
+					Hash and sign JSON using MIMC
 				</p>
+				<p>Using max JSON size: {MAX_JSON_SIZE}</p>
 
-				<pre>Witness Inputs</pre>
+				<p>Witness Inputs</p>
 
-				<textarea 
-					label="json input" rows="4"
+				<input
 					type="text" required={true} value={jsonInput} 
 					onChange={changeJsonInput} placeholder="{'hi':'a'}"
-				></textarea>
+				></input>
 				<br/>
-				<button onClick={handleJsonInput}>Handle Input</button>
+				<button onClick={handleInput}>Handle Input</button>
 				<br/>
+				JSON ascii array input: 
+				<pre>{paddedJsonAscii}</pre>
+				<p>Proof input:</p>
+				<div width="300">
+					<pre>{JSON.stringify(proofInput)}</pre>
+				</div>
 
-				{/* <p>
-					The underlying circuit is from the <a href="https://github.com/iden3/snarkjs">snarkjs readme</a>
-				</p>
-				<pre>Witness Inputs</pre>
-
-				<Field label="Input a:">
-					<Input type="text" required={true} value={a} onChange={changeA} placeholder="e.g. 3" />
-				</Field>
-				<Field label="Input b:">
-					<input type="text" required={true} value={b} onChange={changeB} placeholder="e.g. 11" />
-				</Field> */}
-
-				<button onClick={runProofs}>Generate Proof</button>
+				<button onClick={generateProof}>Generate Proof</button>
 				<br/>
-				Proof: <p>{proof}</p>
-				Signals: <p>{signals}</p>
-				Result:
+				<p>Proof</p>
+				<pre>{proof}</pre>
+				<p>Signals [hash, publickey[0], publickey[1]]</p>
+				<pre>{signals}</pre>
+				<button onClick={handleVerifyProof}>Verify Proof</button>
+				<p>Result:</p>
 				{proof.length > 0 && <p>{isValid ? "Valid proof" : "Invalid proof"}</p>}
 			</header>
 		</div>
