+++ src/zkproof/circuit_1400.circom
@@ -1,0 +1,5 @@
+pragma circom 2.0.0;
+
+include "./circuit.circom";
+
+component main { public [ pubkey ] } = HashAndSignJson(1400); // 515306 constraints
