--- src/zkproof/input.json
@@ -1,1 +1,0 @@
-{"a": 3, "b": 11}
