+++ src/zkproof/input_25.json
@@ -1,0 +1,1 @@
+{"json":[123,34,110,97,109,101,34,58,34,97,98,101,114,107,101,34,125,0,0,0,0,0,0,0,0],"expected_hash":"17714356765695238841646923571006253199580625922870994190188187434035240810052","pubkey":["1891156797631087029347893674931101305929404954783323547727418062433377377293","14780632341277755899330141855966417738975199657954509255716508264496764475094"],"signature_R8x":"16758573265488727828545935640943057777209470912840507105467798116374058104223","signature_R8y":"18392996603700776622368669993552828424883035969510766989506680336244288688829","signature_S":"2437348321815789502892027529529573936225726900639765107898121452273358163588"}
