+++ src/zkproof/proof_1400.json
@@ -1,0 +1,28 @@
+{
+ "pi_a": [
+  "21356266621207522006096580255368837453678907477335086765244483108855339734499",
+  "2666077575513923806189709244778483530933607738512688897718321300117660085100",
+  "1"
+ ],
+ "pi_b": [
+  [
+   "2913810597798163141380858720548949900644091307885319423772275172547207315744",
+   "20598866958146695496828426183513411711405986510906800121478866122166542325634"
+  ],
+  [
+   "12795828725074993868430401827882486210795684051327138010921784586393034043656",
+   "1838108040090111555970506397115894239318319103368163718184274747197950075954"
+  ],
+  [
+   "1",
+   "0"
+  ]
+ ],
+ "pi_c": [
+  "13716750946900064983445511476984209691069277616991177516228255559396201486530",
+  "9528296362474102494826411607954336296863878639958917171183963563455174196289",
+  "1"
+ ],
+ "protocol": "groth16",
+ "curve": "bn128"
+}
