+++ src/zkproof/proof_25.json
@@ -1,0 +1,28 @@
+{
+ "pi_a": [
+  "8846274105609359504967257290984937662759971731440965809960141816224975898921",
+  "18883597238128166522094354187737760343572003061759327052998133121404645611342",
+  "1"
+ ],
+ "pi_b": [
+  [
+   "11810196313199473527681903092683563107840948418919661368273071653086698316772",
+   "16095029422895826452012190342138952301409567554539338053672183404742473926689"
+  ],
+  [
+   "19732130309594002981609153815508853371146257702572528487168968908610811437455",
+   "10984421405630046667120139413519461722024683828139187275853747597515073718086"
+  ],
+  [
+   "1",
+   "0"
+  ]
+ ],
+ "pi_c": [
+  "9051567190087675490087572244669558113573371946000994359430640571896685713740",
+  "9453762522383520830810307315392630803024470635693178434918385869919084478463",
+  "1"
+ ],
+ "protocol": "groth16",
+ "curve": "bn128"
+}
