+++ src/zkproof/public_1400.json
@@ -1,0 +1,5 @@
+[
+ "18156521791552148762641864581372674129058250433775146882183032429707976673106",
+ "1891156797631087029347893674931101305929404954783323547727418062433377377293",
+ "14780632341277755899330141855966417738975199657954509255716508264496764475094"
+]
