+++ src/zkproof/public_25.json
@@ -1,0 +1,5 @@
+[
+ "17714356765695238841646923571006253199580625922870994190188187434035240810052",
+ "1891156797631087029347893674931101305929404954783323547727418062433377377293",
+ "14780632341277755899330141855966417738975199657954509255716508264496764475094"
+]
