--- src/zkproof/verification_key.json
@@ -1,94 +1,0 @@
-{
- "protocol": "groth16",
- "curve": "bn128",
- "nPublic": 1,
- "vk_alpha_1": [
-  "14437500432584531866640327455465178049603581077774177395335838426780673031434",
-  "19802248546963017402741647629546830141571946580593389067654713793747707870342",
-  "1"
- ],
- "vk_beta_2": [
-  [
-   "11010421159950922267767796513605001744363160495046747954018883179573267059774",
-   "8538544126023396852457246166663074302348714932391788072877449778425867364364"
-  ],
-  [
-   "18294248763244114729332962572038917008281538418140377769921143667055656486962",
-   "18511890472759462468392356223385791734833981224383944479009236552601725022689"
-  ],
-  [
-   "1",
-   "0"
-  ]
- ],
- "vk_gamma_2": [
-  [
-   "10857046999023057135944570762232829481370756359578518086990519993285655852781",
-   "11559732032986387107991004021392285783925812861821192530917403151452391805634"
-  ],
-  [
-   "8495653923123431417604973247489272438418190587263600148770280649306958101930",
-   "4082367875863433681332203403145435568316851327593401208105741076214120093531"
-  ],
-  [
-   "1",
-   "0"
-  ]
- ],
- "vk_delta_2": [
-  [
-   "14690287786506416505988263860112384903579673016973724581277995223222420111613",
-   "1859743975662324783071803682311573423447117654704736051200525972711494682816"
-  ],
-  [
-   "17268437707140635525361192089167432367602436331743930421841138940734999971275",
-   "4681056303721876284733851506134339359373419900985129767879283540753741262079"
-  ],
-  [
-   "1",
-   "0"
-  ]
- ],
- "vk_alphabeta_12": [
-  [
-   [
-    "20609006434292023501013068964919213432222044920756299862696167706869752452874",
-    "1033467592195202820879651814677275671653520851809520747909930955448521171265"
-   ],
-   [
-    "9255089529384776445454281874064919385649231323951035231023589768597347835244",
-    "11687642541077894857607849991648466135371840879969002254830057468948087361361"
-   ],
-   [
-    "20948537127598136388075872703726582808407103406878183216178741219208490744120",
-    "20530642616954621101182575049966841343371821965253392898368947243146122080226"
-   ]
-  ],
-  [
-   [
-    "3562941910550177597120870340761877553373335999790615888091544997512368840745",
-    "10281765751857513547654715524342514635900691828890056126819859812090315417832"
-   ],
-   [
-    "3097716122812015301853078003971036298235018384478095339815840212144736125339",
-    "20805179056480032827312442937039167330051774638349699967683921398560934448537"
-   ],
-   [
-    "6730041892790237060916390851523014903586654994590293076271294468990335515059",
-    "2114367023745732527606492979644017137803784265989454163254030419283490095456"
-   ]
-  ]
- ],
- "IC": [
-  [
-   "15901306312448524246355652168759400662031931686483030635148528092203146922533",
-   "4569403428323027289162756641287378261271796105032474946684166815102195790945",
-   "1"
-  ],
-  [
-   "19971067590538856074687329403689801712808085278154604900241767506501432765419",
-   "6410715836171018193796590852102204928308356428670371844084574646204241415210",
-   "1"
-  ]
- ]
-}
