+++ src/zkproof/verification_key_1400.json
@@ -1,0 +1,104 @@
+{
+ "protocol": "groth16",
+ "curve": "bn128",
+ "nPublic": 3,
+ "vk_alpha_1": [
+  "11531318685811437041426321596256588195302899808334830443733461404527243609379",
+  "16265228672050288363916744427266526669390524501967543288600732344560885579632",
+  "1"
+ ],
+ "vk_beta_2": [
+  [
+   "10299481726648418690905729745496735235633633833092928217084906259233487207655",
+   "1327017437584537555887439335139122788208076993892276583852415314922291033166"
+  ],
+  [
+   "9139760378194901119934708685349268354274802358648860814208946668426435724320",
+   "14990542748237955205559810909058398691878554704989493631922857305356505033969"
+  ],
+  [
+   "1",
+   "0"
+  ]
+ ],
+ "vk_gamma_2": [
+  [
+   "10857046999023057135944570762232829481370756359578518086990519993285655852781",
+   "11559732032986387107991004021392285783925812861821192530917403151452391805634"
+  ],
+  [
+   "8495653923123431417604973247489272438418190587263600148770280649306958101930",
+   "4082367875863433681332203403145435568316851327593401208105741076214120093531"
+  ],
+  [
+   "1",
+   "0"
+  ]
+ ],
+ "vk_delta_2": [
+  [
+   "10857046999023057135944570762232829481370756359578518086990519993285655852781",
+   "11559732032986387107991004021392285783925812861821192530917403151452391805634"
+  ],
+  [
+   "8495653923123431417604973247489272438418190587263600148770280649306958101930",
+   "4082367875863433681332203403145435568316851327593401208105741076214120093531"
+  ],
+  [
+   "1",
+   "0"
+  ]
+ ],
+ "vk_alphabeta_12": [
+  [
+   [
+    "7485030149434230845685995683199563142133379004815949639343129449430141956343",
+    "16291547097695579368773311410462916205670029121222626790360534516623586330852"
+   ],
+   [
+    "233697959620407485223693166450914185935341906514829547337582227443726528918",
+    "3339027281052550997883207072944566154939983487741983508045782567201708700825"
+   ],
+   [
+    "11948763297521507018881182700798464083980220249503882633157937461553747046274",
+    "12828092003247986151696566985129474775871290154169181846054926167883996848654"
+   ]
+  ],
+  [
+   [
+    "12174994996607423004621146004149394221081380137797249589862162209251051660737",
+    "923459360724068983532255292975475242753718063202029273213721585822800680040"
+   ],
+   [
+    "6191251288116380150634455611077246591654887879953898757854075006755855824372",
+    "5182584784765022848429570386582285325856865234947423516262788904055009709229"
+   ],
+   [
+    "7682542101149980951193493695413267842669021851485538995674716200465571361931",
+    "9346124329208572098263133182166946902725272986218931035274846430485419136485"
+   ]
+  ]
+ ],
+ "IC": [
+  [
+   "492344413894655031255969005471052097812288289152575784676699348906963922132",
+   "19424863821043916748699423252640144537983629512385853599925470228368683393404",
+   "1"
+  ],
+  [
+   "6057670345158976834961288862246959081737044372579771775478422122646014538500",
+   "7469474712534086365002424772722434333706102279174907039675205204996069079785",
+   "1"
+  ],
+  [
+   "1811096042408190537290666006455389191088014069139693746088159747493283966285",
+   "2039924313867901408630040134299500295611027223543711426749357714321365358341",
+   "1"
+  ],
+  [
+   "17749120307025257868010575033046553577552774915154366589217586822062982986541",
+   "5966878171735984831908145149610888248474543644025207595861403780352091818702",
+   "1"
+  ]
+ ]
+}
