+++ src/zkproof/verification_key_25.json
@@ -1,0 +1,104 @@
+{
+ "protocol": "groth16",
+ "curve": "bn128",
+ "nPublic": 3,
+ "vk_alpha_1": [
+  "20998223588842477229106979951757629899531543025347176409340636768225656463775",
+  "6746241957381828750667639020116525551795474930949003380043924072184110794394",
+  "1"
+ ],
+ "vk_beta_2": [
+  [
+   "20289546218581188778025360951578800878484849614134770948019722518303202074415",
+   "7411952280561234651707573615800043255577838064353912674819738113219148076168"
+  ],
+  [
+   "646477522419467601084510726532234236631308005125190143891321769071107751354",
+   "21730703263285007381672169231413962516817608500134029835614864488534147372199"
+  ],
+  [
+   "1",
+   "0"
+  ]
+ ],
+ "vk_gamma_2": [
+  [
+   "10857046999023057135944570762232829481370756359578518086990519993285655852781",
+   "11559732032986387107991004021392285783925812861821192530917403151452391805634"
+  ],
+  [
+   "8495653923123431417604973247489272438418190587263600148770280649306958101930",
+   "4082367875863433681332203403145435568316851327593401208105741076214120093531"
+  ],
+  [
+   "1",
+   "0"
+  ]
+ ],
+ "vk_delta_2": [
+  [
+   "10857046999023057135944570762232829481370756359578518086990519993285655852781",
+   "11559732032986387107991004021392285783925812861821192530917403151452391805634"
+  ],
+  [
+   "8495653923123431417604973247489272438418190587263600148770280649306958101930",
+   "4082367875863433681332203403145435568316851327593401208105741076214120093531"
+  ],
+  [
+   "1",
+   "0"
+  ]
+ ],
+ "vk_alphabeta_12": [
+  [
+   [
+    "11162841316347643609099885430598622135422969922127060595021615015751829563883",
+    "21302436828589844833137562434476740830835075159675910089071172247916247515108"
+   ],
+   [
+    "644208081507865462081179660166187654968526038280737173269895341254851188283",
+    "13925582108735023166011287402824873971704854271726103995072101840845212996147"
+   ],
+   [
+    "20400089943119055684100529203040168070370357351102529613745105016386886777000",
+    "13584572253412288810809242134854688495917140857968514290619623445597639548719"
+   ]
+  ],
+  [
+   [
+    "14963811627294030431633176280292679159559472933799286113923485542377630299089",
+    "6040255866130499520858582611130108656171696503008994037920193014753330542081"
+   ],
+   [
+    "9361800230770374508888467923405746768747224227683864937304235929266621883319",
+    "15628366752354810791158799130499475885859585745044090987493654141211268080315"
+   ],
+   [
+    "16126735324119875890772150514247655495477533394104482926360625682437568678651",
+    "19944255216942502732453893835259629119966942989139308287764658908604797670108"
+   ]
+  ]
+ ],
+ "IC": [
+  [
+   "9954513244771099486857334343997931404216706324571824420629990492938134453315",
+   "15394675356814203732359481311459208201745889012156795124733941119424152307973",
+   "1"
+  ],
+  [
+   "6592638334689437800932792736212587921301222393908811440480753751469250061743",
+   "15456827322752700288275748790820159748950626496427762078663977161177468335070",
+   "1"
+  ],
+  [
+   "8373658268576280376179676639655939097449032664172080744433372467469907519967",
+   "16097938068264078482934813956045208255625296533496091428680031667094941940213",
+   "1"
+  ],
+  [
+   "1387963509513713876864448957701057776791116937431046286078718699778874149132",
+   "7761420061908348756966309215066461135266798054409237716868337027477542732191",
+   "1"
+  ]
+ ]
+}
